Add a toAff' for native code that does not fail with Errors or strings
For JavaScript code that does not reject promises with an Error or a
string, users of the toAff API lose tha ability to have meaningful
errors. This adds a new API that allows the user of the API to provide a
custom conversion function that takes the reject value as a Foreign and
provides an Error object instead.